fixed styles

+++ src/main.cpp
@@ -1,11 +1,10 @@
-
-#include "fibonacci.h"
 #include <iostream>
+#include "../include/fibonacci.h"
 
 int main(int argc, char const *argv[]) {
   std::string s(argv[1]);
   unsigned int N = std::stoi(s);
-  long long x = fibonacci(N);
+  uint64_t x = fibonacci(N);
   std::cout << "The " << N << "th Fibonacci number is: " << x << std::endl;
   return 0;
 }
